requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
accelerate==1.1.1
diffusers==0.31.0
huggingface-hub==0.26.2
numpy==2.1.3
pillow==11.0.0
safetensors==0.4.5
torch>=2.1.2
torchvision>=0.16.2
torchaudio>=2.1.2
transformers==4.46.2
transparent-background==1.3.3
